+++ nextcloudappstore/user/forms.py
@@ -1,7 +1,7 @@
 from allauth.account.utils import filter_users_by_email
 from django import forms
 from django.contrib.auth import get_user_model
-from django.forms import EmailField, CharField
+from django.forms import EmailField, CharField, PasswordInput
 from django.utils.translation import ugettext_lazy as _
 from snowpenguin.django.recaptcha2.fields import ReCaptchaField
 from snowpenguin.django.recaptcha2.widgets import ReCaptchaWidget
@@ -36,16 +36,30 @@ class DeleteAccountForm(forms.Form):
 
 
 class AccountForm(forms.ModelForm):
+    passwd = CharField(widget=PasswordInput(), label=_('Confirm password'),
+                       help_text=_('Password is required to prevent '
+                                   'unauthorized users from changing your '
+                                   'email address and resetting your '
+                                   'password. This field does not update your '
+                                   'password!'))
+
     class Meta:
         model = get_user_model()
         fields = ('first_name', 'last_name', 'email')
 
     def clean_email(self):
-        value = self.cleaned_data["email"]
+        value = self.cleaned_data['email']
         users = filter_users_by_email(value)
         if [u for u in users if u.pk != self.instance.pk]:
             msg = _(
-                "This e-mail address is already associated with another "
-                "account.")
+                'This e-mail address is already associated with another '
+                'account.')
             raise forms.ValidationError(msg)
         return value
+
+    def clean_passwd(self):
+        value = self.cleaned_data['passwd']
+        if self.instance.check_password(value):
+            return value
+        else:
+            raise forms.ValidationError(_('Invalid password'))